modem: simcom: driver abstraction for simcom
add support for simcom Y7080 and abstracted with simcom
sim7080
