+++ drivers/modem/CMakeLists.txt
@@ -37,7 +37,14 @@ if (CONFIG_MODEM_HL7800)
   zephyr_library_sources(hl7800.c)
 endif()
 
-if (CONFIG_MODEM_SIM7080)
+if (CONFIG_MODEM_SIMCOM_SIM7080)
 	zephyr_library_include_directories(${ZEPHYR_BASE}/subsys/net/ip)
 	zephyr_library_sources(simcom-sim7080.c)
+	zephyr_library_sources(simcom-xxxx.c)
+endif()
+
+if (CONFIG_MODEM_SIMCOM_Y7080)
+	zephyr_library_include_directories(${ZEPHYR_BASE}/subsys/net/ip)
+	zephyr_library_sources(simcom-y7080.c)
+	zephyr_library_sources(simcom-xxxx.c)
 endif()
