+++ drivers/modem/Kconfig.simcom-xxxx
@@ -1,0 +1,129 @@
+# Simcom Modem driver options
+
+# Copyright (C) 2021 metraTec GmbH
+# Copyright (C) 2023 Linumiz
+# SPDX-License-Identifier: Apache-2.0
+
+menuconfig MODEM_SIMCOM
+	bool "Simcom Modem Driver for 7080 series"
+	select MODEM_CONTEXT
+	select MODEM_CMD_HANDLER
+	select MODEM_IFACE_UART
+	select MODEM_SOCKET
+	select NET_OFFLOAD
+	select NET_SOCKETS_OFFLOAD
+	imply GPIO
+	help
+	  Enables the driver for the Simcom modem for 7080 series.
+
+if MODEM_SIMCOM
+
+choice SIMCOM_MODEM_SELECTION
+	prompt "Simcom modem selection"
+	default MODEM_SIMCOM_SIM7080
+	help
+	 this helps in selecting simcom modem.
+
+config MODEM_SIMCOM_SIM7080
+	bool "simcom sim7080 modem"
+	help
+	  Enables the driver for the Simcom sim7080 modem.
+
+config MODEM_SIMCOM_Y7080
+	bool "simcom Y7080 modem"
+	help
+	  Enables the driver for the Simcom Y7080 modem.
+
+endchoice
+
+config MODEM_SIMCOM_RX_STACK_SIZE
+	int "Stack size for the simcom modem driver rx thread"
+	default 2048 if MODEM_SIMCOM_Y7080
+	default 1028 if MODEM_SIMCOM_SIM7080
+	help
+	  This stack is used by the simcom RX thread.
+
+config MODEM_SIMCOM_RX_WORKQ_STACK_SIZE
+	int "Stack size for the simcom modem driver work queue"
+	default 2048
+	help
+	 This stack is used by the work queue.
+
+config MODEM_SIMCOM_INIT_PRIORITY
+	int "simcom driver init priority"
+	default 80
+	help
+	  Simcom driver initialization priority.
+
+config MODEM_SIMCOM_LTE_BANDS
+	string "LTE bands the driver can use"
+	default "8,20,28"
+	help
+	 Comma separated list of usable lte bands.
+
+config MODEM_SIMCOM_PSM_EDRX_DISABLE
+	bool "Disable PSM EDRX"
+	default n
+	help
+	 Disable the PSM and EDRX mode.
+
+config MODEM_SIMCOM_APN
+	string "APN for establishing a network connection"
+	default "internet"
+	help
+	  This setting is used to set the APN name for the network connection
+	  context. This value is specific to the network provider and may
+	  need to be changed.
+
+choice MODEM_SIMCOM_RAT
+	bool "Radio Access Technology Mode"
+	default MODEM_SIMCOM_RAT_NB1
+
+config MODEM_SIMCOM_RAT_NB1
+	bool "NB-IoT"
+	help
+	 Enable LTE NB-IoT mode.
+
+config MODEM_SIMCOM_RAT_M1
+	bool "Cat-M1"
+	help
+	 Enable Cat-M1 mode.
+
+config MODEM_SIMCOM_RAT_GSM
+	bool "GSM"
+	help
+	 Enable GSM mode.
+
+endchoice
+
+choice SIMCOM_GNSS
+	bool "SIMCOM GNSS"
+	default SIMCOM_GNSS_COLD_START
+
+config SIMCOM_GNSS_WARM_START
+	bool "GNSS warm start"
+	help
+	 Enable gnss warm start mode.
+
+config SIMCOM_GNSS_HOT_START
+	bool "GNSS Hot start"
+	help
+	 Enable gns hot start mode.
+
+config SIMCOM_GNSS_COLD_START
+	bool "GNSS cold start"
+	help
+	 Enable gnss cold start mode.
+
+endchoice
+
+if MODEM_SIMCOM_SIM7080
+
+config MODEM_SIMCOM_FTP_AND_SMS
+	bool "FTP and SMS"
+	help
+	 Enable FTP and SMS support.
+
+endif
+
+endif
