+++ dts/bindings/modem/simcom,sim7080.yaml
@@ -1,13 +1,9 @@
 # Copyright (C) 2021 metraTec GmbH
+# Copyright (C) 2023 Linumiz
 # SPDX-License-Identifier: Apache-2.0
 
-description: Simcom Sim7080 modem
+description: Simcom SIM7080 Modem
 
 compatible: "simcom,sim7080"
 
-include: uart-device.yaml
-
-properties:
-  mdm-power-gpios:
-    type: phandle-array
-    required: true
+include: simcom,7080.yaml
