+++ dts/bindings/modem/simcom,y7080.yaml
@@ -1,0 +1,8 @@
+# Copyright (C) 2023 Linumiz
+# SPDX-License-Identifier: Apache-2.0
+
+description: Simcom Y7080 Modem
+
+compatible: "simcom,y7080"
+
+include: simcom,7080.yaml
