+++ dts/bindings/modem/simcom.yaml
@@ -1,0 +1,13 @@
+# Copyright (C) 2023 Linumiz
+# SPDX-License-Identifier: Apache-2.0
+
+description: Simcom modem 7080 series
+
+compatible: "simcom,7080"
+
+include: uart-device.yaml
+
+properties:
+  mdm-power-gpios:
+    type: phandle-array
+    required: true
